backend/requirements.txt
Flask==2.2.5
Flask-JWT-Extended==4.5.2
Flask-SQLAlchemy==3.0.3
Flask-Cors==6.0.1
bcrypt==4.0.1
qrcode==7.4.2
pillow==11.3.0
python-dotenv==1.0.0
Werkzeug==2.2.3
# gunicorn is only needed for deployment
# gunicorn==21.2.0
